+++ CHANGELOG.md
@@ -16,11 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - add a new parameter `insuranceFundArg` to `initialize` of ClearingHouse
+- add a new parameter `orderBookArg` to `initialize` of AccountBalance
 
 ### Removed
 
 - `AccountBalance.getNetQuoteBalanceAndPendingFee`
 - remove parameter `insuranceFundArg` from `initialize` of Exchange
+- remove parameter `exchangeArg` from `initialize` of AccountBalance
 
 ## [0.14.0-staging] - 2021-11-17